Lronac2pds App and Test Update (#4135)
* Converted lronac2pds to new application format

* Updated tests for lronac2pds to gtests

* Removed old lronac2pds test makefiles

* Removed local path

* Addressed PR comments
