--- isis/src/lro/apps/lronac2pds/tsts/Makefile
@@ -1,4 +1,0 @@
-BLANKS = "%-6s"    
-LENGTH = "%-40s"
-
-include $(ISISROOT)/make/isismake.tststree
